requirements.txt
flask
flask-cors
rembg
onnxruntime
pillow
gunicorn
pymongo
mysql-connector-python
psycopg2-binary
